Changed structure, fixed imports
