--- bot/bot.py
@@ -1,34 +1,0 @@
-import asyncio
-import logging
-import sys
-
-from os import getenv
-from dotenv import load_dotenv
-
-from aiogram import Bot, Dispatcher
-from aiogram.enums import ParseMode
-from aiogram.client.session.aiohttp import AiohttpSession
-from aiogram.client.bot import DefaultBotProperties
-
-from bot.handlers import router
-
-load_dotenv()
-
-TG_TOKEN = getenv('TOKEN')
-bot = Bot(
-        token=TG_TOKEN,
-        session=AiohttpSession(),
-        default=DefaultBotProperties(parse_mode=ParseMode.HTML)
-    )
-
-dp = Dispatcher()
-dp.include_router(router)
-
-
-async def main() -> None:
-    await dp.start_polling(bot)
-
-
-if __name__ == "__main__":
-    logging.basicConfig(level=logging.INFO, stream=sys.stdout)
-    asyncio.run(main())
